Replace @Generated annotation with comment
Since Java 9 the @Generated annotation is not contained in the JRE by
default anymore and an additional dependency is necessary.
To not force the user of a generated Facade to introduce such a
dependency, the annotation was now converted into a comment.

In the future it would be possible to decide between the command and the
real annotation via commandline switch. This would allow the user to
decide whether to use the annotation or not.