Replace java globbing with ant-style globbing again
Java globbing uses an uncommon syntax for 0-n directory hierarchies:
  **.properties
instead of the more common
  **/*.properties
Also it doesn't detect if equivalent paths are the same due to some
indirection.
Therefore plexus-utils is used again to do the matching againts
patterns. A bit custom logic around that allows normalization of patterns
and paths and specifying patterns either as absolute paths or as
relative to the propertyRootDirectory.