Replace plexus DirectoryScanner with custom implementation
This implementation allows specifying i18nIncludes/i18nExcludes as
relative to their propertyRootDirectory or as a path including the
propertyRootDirectory (which makes it easier to specify them on command
line via command line completion).

It is still not perfect, since it utilizes the globbing feature of
Java 7 and that differs in behaviour from the typical
posix/bash/zsh/fish globbing.  Especially '**/*.properties' does not
work, because it /requires/ the properties to be in a subdirectory.
Usually a user would expect that this glob matches properties file in
any subdirectory hierarchy as well in the directory /without/ an
subdirectory.