Provide user guide for all releases
An index.html file in the root directory lists all available user
guides. This one is an asciidoc file that gets generated to
HTML.¹

It contains a link to each releases docu.
It contains a linke for the current git HEAD.
It contains a link for the latest release version

On releasing Kilt

 - a new directory for that release is created with that user guide
 - the latest link will be updated to point to the new release
 - the index.adoc gets a new entry for the new release

[¹] Problem: Where to store the index.adoc? In the docs folder?