Use NullAway to check for problems with null values
The maven plugin, the ant task and the CLI get their values through
initializers. It is not fully possible to assert non-values there.
Therefore we assume these are correctly initialized.

The `reorder` task needs the `template` parameter if `byKey` is not set.
Even though this is checked in an additional validate method NullAway is
not aware of this. Therefore an additional null-check is done now where
`template` is used.

Charset handling has been unified. A charset is now required. It will
default to UTF-8 if not given.

Since using errorprone (which is used by NullAway) requires a different
configuration to be used with Java 8, the build now requires Java 9+.
The target release is still Java 8, so Kilt still can be used with Java 8.