Try to build native image
This doesn’t work yet.

Building for Windows fails with:
“Failed to execute goal de.poiu.kilt:kilt-maven-plugin:1.0.2-SNAPSHOT:create-facade (kilt-facade-generation) on project kilt-runtime: Execution kilt-facade-generation of goal de.poiu.kilt:kilt-maven-plugin:1.0.2-SNAPSHOT:create-facade failed: Illegal char <*> at index 0: **/*.properties”

Building for Linux with musl (as a statically linked image) fails with:
“com.oracle.svm.core.util.VMError$HostedError: Unable to run
'/tmp/SVM-5801301810423631605/AArch64LibCHelperDirectives' to compute
offsets in C data structures. Is it possible that your antivirus
software interferes and puts the resulting file into quarantine?”

Building for Linux without musl succeeds, but generates an binary that
fails at runtime with the message:
“The current machine does not support all of the following CPU features that are required by the image: [CX8, CMOV, FXSR, MMX, SSE, SSE2, SSE3, SSSE3, SSE4_1, SSE4_2, POPCNT, LZCNT, AVX, AVX2, BMI1, BMI2, FMA].
Please rebuild the executable with an appropriate setting of the -march option.”